Move constants and helper functions to processor/utils
Although created for tests, these functions and values will be used by the
transaction handler. They've been moved to a root level where they can be
shared.
