+++ code/part-two/processor/handler.js
@@ -2,18 +2,13 @@
 
 const { createHash } = require('crypto');
 const { TransactionHandler } = require('sawtooth-sdk/processor/handler');
+const { FAMILY_NAME, FAMILY_VERSION, NAMESPACE } = require('./utils/constants');
 
-const getAddress = (key, length) => {
-  return createHash('sha512').update(key).digest('hex').slice(0, length);
-};
-
-const FAMILY = 'cryptomoji';
-const NAMESPACE = getAddress(FAMILY, 6);
 
 class MojiHandler extends TransactionHandler {
   constructor () {
     console.log('Initializing cryptomoji handler with namespace:', NAMESPACE);
-    super(FAMILY, ['1.0'], [NAMESPACE]);
+    super(FAMILY_NAME, [ FAMILY_VERSION ], [ NAMESPACE ]);
   }
 
   apply (txn, context) {
